+++ extensions/positron-code-cells/src/folding.ts
@@ -2,7 +2,7 @@
  *  Copyright (C) 2023 Posit Software, PBC. All rights reserved.
  *--------------------------------------------------------------------------------------------*/
 import * as vscode from 'vscode';
-import { generateCellRangesFromDocument } from './codeLenseProvider';
+import { generateCellRangesFromDocument } from './codeLenses';
 
 export function registerFoldingRangeProvider(context: vscode.ExtensionContext): void {
 	context.subscriptions.push(
